requirements.txt
# Deep Learning & AI
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Computer Vision & Image Processing
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0

# Diffusion Models & Transformers
diffusers>=0.20.0
transformers>=4.21.0

# Web Framework (Django)
Django>=4.2.0

# Data & Datasets
datasets>=2.0.0
requests>=2.28.0

# Visualization & Progress
matplotlib>=3.5.0
tqdm>=4.64.0

# Utilities
typing-extensions>=4.0.0